Added a nice big LoginButton
Hooks into the AuthenticationContext we created. Must be located somewhere under an AuthenticationContextProvider to work.

When there is no good authentication, the button is a Log-In button that shows a dialog and calls Authentication.login(...)
When there *is* an authentication already, the button is a Log-Out button that clears the authentication-info.