Added an App-Bar and a Login/Logout Menu
Added an AppNavBar, to go at the top of the screen.

It contains a LoginButtonMenu, which presents a button, associated with a popup menu. That popup menu can in turn trigger a LoginDialog.

LoginDialog presents a LoginForm, as usual, but doesn't handle any of the submission-logic. That must be handled by the caller.